Enable startup error-reporting on Mac
We attach Eto's crash handler as first order of business in Main().
This means that if an error occurs after the main Form is constructed
and Eto initialises itself (attaching the crash handler again), its
stack trace will show up twice in the crash report, but that's a good
tradeoff in my eyes.